+++ tests/codegen_refs/sum_broadcast.c
@@ -1,11 +1,12 @@
 #include <stdint.h>
+#include <stddef.h>
 #include <stdbool.h>
 #include "ops_scalar_f32.h"
 
 void node1_add_f32(const float a[2][3][4], const float b[1][4], float out[2][3][4]) {
-    for (int64_t i0 = 0; i0 < 2; ++i0) {
-        for (int64_t i1 = 0; i1 < 3; ++i1) {
-            for (int64_t i2 = 0; i2 < 4; ++i2) {
+    for (size_t i0 = 0; i0 < 2; ++i0) {
+        for (size_t i1 = 0; i1 < 3; ++i1) {
+            for (size_t i2 = 0; i2 < 4; ++i2) {
                 out[i0][i1][i2] = ref_scalar_f32_add(a[i0][i1][i2], b[0][i2]);
             }
         }
@@ -13,10 +14,10 @@ void node1_add_f32(const float a[2][3][4], const float b[1][4], float out[2][3][
 }
 
 void node2_sum_f32(const float a[2][3][4], float out[2][4]) {
-    for (int64_t i0 = 0; i0 < 2; ++i0) {
-        for (int64_t i1 = 0; i1 < 4; ++i1) {
+    for (size_t i0 = 0; i0 < 2; ++i0) {
+        for (size_t i1 = 0; i1 < 4; ++i1) {
             float acc = 0.0f;
-            for (int64_t r1 = 0; r1 < 3; ++r1) {
+            for (size_t r1 = 0; r1 < 3; ++r1) {
                 acc += a[i0][r1][i1];
             }
             out[i0][i1] = acc;
